chore: relocate CMAP code to `lib/cmap`
