+++ lib/cmap/connection.js
@@ -2,12 +2,12 @@
 
 const EventEmitter = require('events');
 const MessageStream = require('./message_stream');
-const MongoError = require('../error').MongoError;
-const MongoWriteConcernError = require('../error').MongoWriteConcernError;
-const wp = require('../wireprotocol');
-const apm = require('../connection/apm');
-const updateSessionFromResponse = require('../sessions').updateSessionFromResponse;
-const uuidV4 = require('../utils').uuidV4;
+const MongoError = require('../core/error').MongoError;
+const MongoWriteConcernError = require('../core/error').MongoWriteConcernError;
+const wp = require('../core/wireprotocol');
+const apm = require('../core/connection/apm');
+const updateSessionFromResponse = require('../core/sessions').updateSessionFromResponse;
+const uuidV4 = require('../core/utils').uuidV4;
 
 const kStream = Symbol('stream');
 const kQueue = Symbol('queue');
@@ -276,7 +276,6 @@ class CMAPConnection {
 
   destroy() {}
 }
-
 
 module.exports = {
   Connection,
