+++ lib/cmap/connection_manager.js
