+++ lib/cmap/counter.js
