+++ lib/cmap/errors.js
