+++ lib/cmap/events.js
