+++ lib/cmap/index.js
@@ -1,0 +1,5 @@
+'use strict';
+
+const ConnectionPool = require('./connection_pool').ConnectionPool;
+
+module.exports = { ConnectionPool };
