+++ lib/cmap/message_stream.js
@@ -2,20 +2,20 @@
 
 const Duplex = require('stream').Duplex;
 const BufferList = require('bl');
-const MongoParseError = require('../error').MongoParseError;
-const decompress = require('../wireprotocol/compression').decompress;
-const Response = require('../connection/commands').Response;
-const BinMsg = require('../connection/msg').BinMsg;
-const MongoError = require('../error').MongoError;
-const OP_COMPRESSED = require('../wireprotocol/shared').opcodes.OP_COMPRESSED;
-const OP_MSG = require('../wireprotocol/shared').opcodes.OP_MSG;
-const MESSAGE_HEADER_SIZE = require('../wireprotocol/shared').MESSAGE_HEADER_SIZE;
-const COMPRESSION_DETAILS_SIZE = require('../wireprotocol/shared').COMPRESSION_DETAILS_SIZE;
-const opcodes = require('../wireprotocol/shared').opcodes;
-const compress = require('../wireprotocol/compression').compress;
-const compressorIDs = require('../wireprotocol/compression').compressorIDs;
-const uncompressibleCommands = require('../wireprotocol/compression').uncompressibleCommands;
-const Msg = require('../connection/msg').Msg;
+const MongoParseError = require('../core/error').MongoParseError;
+const decompress = require('../core/wireprotocol/compression').decompress;
+const Response = require('../core/connection/commands').Response;
+const BinMsg = require('../core/connection/msg').BinMsg;
+const MongoError = require('../core/error').MongoError;
+const OP_COMPRESSED = require('../core/wireprotocol/shared').opcodes.OP_COMPRESSED;
+const OP_MSG = require('../core/wireprotocol/shared').opcodes.OP_MSG;
+const MESSAGE_HEADER_SIZE = require('../core/wireprotocol/shared').MESSAGE_HEADER_SIZE;
+const COMPRESSION_DETAILS_SIZE = require('../core/wireprotocol/shared').COMPRESSION_DETAILS_SIZE;
+const opcodes = require('../core/wireprotocol/shared').opcodes;
+const compress = require('../core/wireprotocol/compression').compress;
+const compressorIDs = require('../core/wireprotocol/compression').compressorIDs;
+const uncompressibleCommands = require('../core/wireprotocol/compression').uncompressibleCommands;
+const Msg = require('../core/connection/msg').Msg;
 
 const kDefaultMaxBsonMessageSize = 1024 * 1024 * 16 * 4;
 const kBuffer = Symbol('buffer');
