--- lib/core/cmap/index.js
@@ -1,5 +1,0 @@
-'use strict';
-
-const ConnectionPool = require('./pool').ConnectionPool;
-
-module.exports = { ConnectionPool };
